An App for the proof of Concept
Android App:
The User says the command and the app captures those commands sends it
to Google's speech to text converter and gets back a string equivalent
of the voice command and send it to the PC over wifi.

PC:
There will be a server running on the PC which receives this string and
there will be pre-defined shell scripts for each functions, the server
just maps the command string to the appropriate shell script.